@@ -1,18 +1,12 @@
-const standard = require('@grafana/toolkit/src/config/jest.plugin.config');
-const globals = standard.jestConfig().globals;
-const tsJest = globals['ts-jest'];
+const { grafanaESModules, nodeModulesToTransform } = require('./.config/jest/utils');
+
+// force timezone to UTC to allow tests to work regardless of local timezone
+// generally used by snapshots, but can affect specific tests
+process.env.TZ = 'UTC';
+const originalConfig = require('./.config/jest.config');
 
 module.exports = {
-  ...{
-    ...standard.jestConfig(),
-    setupFilesAfterEnv: ['<rootDir>/src/tests.ts'],
-    modulePathIgnorePatterns: ['<rootDir>/src/module.ts', '<rootDir>/src/selectors.ts', '<rootDir>/src/styles.ts', '<rootDir>/src/types.ts'],
-    globals: {
-      ...globals,
-      'ts-jest': {
-        isolatedModules: tsJest.isolatedModules,
-        tsconfig: tsJest.tsConfig,
-      },
-    },
-  },
+  // Jest configuration provided by Grafana scaffolding
+  ...originalConfig,
+  transformIgnorePatterns: originalConfig.transformIgnorePatterns.map((pattern) => (pattern.startsWith('node_modules') ? `../../${pattern}` : pattern)),
 };
