@@ -1,17 +1,20 @@
 package infinity
 
 import (
-	"fmt"
-	"strings"
+	"context"
+	"errors"
 
+	"github.com/grafana/grafana-infinity-datasource/pkg/models"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/tracing"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/yesoreyeram/grafana-framer/csvFramer"
-	"github.com/yesoreyeram/grafana-framer/gframer"
-	querySrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/query"
+	"github.com/grafana/infinity-libs/lib/go/csvframer"
+	"github.com/grafana/infinity-libs/lib/go/gframer"
 )
 
-func GetCSVBackendResponse(responseString string, query querySrv.Query) (*data.Frame, error) {
+func GetCSVBackendResponse(ctx context.Context, responseString string, query models.Query) (*data.Frame, error) {
+	_, span := tracing.DefaultTracer().Start(ctx, "GetCSVBackendResponse")
+	defer span.End()
 	frame := GetDummyFrame(query)
 	columns := []gframer.ColumnSelector{}
 	for _, c := range query.Columns {
@@ -22,7 +25,7 @@ func GetCSVBackendResponse(responseString string, query querySrv.Query) (*data.F
 			TimeFormat: c.TimeStampFormat,
 		})
 	}
-	csvOptions := csvFramer.CSVFramerOptions{
+	csvOptions := csvframer.FramerOptions{
 		FrameName:          query.RefID,
 		Columns:            columns,
 		Comment:            query.CSVOptions.Comment,
@@ -36,44 +39,18 @@ func GetCSVBackendResponse(responseString string, query querySrv.Query) (*data.F
 	if query.CSVOptions.Columns == "-" || query.CSVOptions.Columns == "none" {
 		csvOptions.NoHeaders = true
 	}
-	if query.Type == querySrv.QueryTypeTSV {
+	if query.Type == models.QueryTypeTSV {
 		csvOptions.Delimiter = "\t"
 	}
-	newFrame, err := csvFramer.CsvStringToFrame(responseString, csvOptions)
-	frame.Meta = &data.FrameMeta{
-		Custom: &CustomMeta{
-			Query: query,
-		},
-	}
-	if err != nil {
-		backend.Logger.Error("error getting response for query", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{
-			Query: query,
-			Error: err.Error(),
-		}
-		return frame, err
-	}
+	newFrame, err := csvframer.ToFrame(responseString, csvOptions)
 	if newFrame != nil {
 		frame.Fields = append(frame.Fields, newFrame.Fields...)
 	}
-	frame, err = GetFrameWithComputedColumns(frame, query.ComputedColumns)
 	if err != nil {
-		backend.Logger.Error("error getting computed column", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
-		return frame, err
-	}
-	frame, err = ApplyFilter(frame, query.FilterExpression)
-	if err != nil {
-		backend.Logger.Error("error applying filter", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
-		return frame, fmt.Errorf("error applying filter. %w", err)
-	}
-	if strings.TrimSpace(query.SummarizeExpression) != "" {
-		return GetSummaryFrame(frame, query.SummarizeExpression, query.SummarizeBy)
-	}
-	if query.Format == "timeseries" && frame.TimeSeriesSchema().Type == data.TimeSeriesTypeLong {
-		if wFrame, err := data.LongToWide(frame, &data.FillMissing{Mode: data.FillModeNull}); err == nil {
-			return wFrame, err
+		if errors.Is(err, csvframer.ErrEmptyCsv) || errors.Is(err, csvframer.ErrReadingCsvResponse) {
+			err = backend.DownstreamError(err)
+		} else {
+			err = backend.PluginError(err)
 		}
 	}
 	return frame, err
