@@ -2,14 +2,17 @@ package infinity
 
 import (
 	"bytes"
+	"context"
 	"encoding/base64"
 	"fmt"
 	"mime/multipart"
 	"net/http"
 	"strings"
 
-	querySrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/query"
-	settingsSrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/settings"
+	"github.com/grafana/grafana-infinity-datasource/pkg/models"
+	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/propagation"
 )
 
 const dummyHeader = "xxxxxxxx"
@@ -20,26 +23,27 @@ const (
 )
 
 const (
-	headerKeyAccept        = "Accept"
-	headerKeyContentType   = "Content-Type"
-	headerKeyAuthorization = "Authorization"
-	headerKeyIdToken       = "X-ID-Token"
+	headerKeyAccept         = "Accept"
+	headerKeyContentType    = "Content-Type"
+	headerKeyAcceptEncoding = "Accept-Encoding"
+	HeaderKeyAuthorization  = "Authorization"
+	HeaderKeyIdToken        = "X-Id-Token"
 )
 
-func ApplyAcceptHeader(query querySrv.Query, settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
-	if query.Type == querySrv.QueryTypeJSON || query.Type == querySrv.QueryTypeGraphQL {
+func ApplyAcceptHeader(_ context.Context, query models.Query, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+	if query.Type == models.QueryTypeJSON || query.Type == models.QueryTypeGraphQL {
 		req.Header.Set(headerKeyAccept, `application/json;q=0.9,text/plain`)
 	}
-	if query.Type == querySrv.QueryTypeCSV {
+	if query.Type == models.QueryTypeCSV {
 		req.Header.Set(headerKeyAccept, `text/csv; charset=utf-8`)
 	}
-	if query.Type == querySrv.QueryTypeXML {
+	if query.Type == models.QueryTypeXML {
 		req.Header.Set(headerKeyAccept, `text/xml;q=0.9,text/plain`)
 	}
 	return req
 }
 
-func ApplyContentTypeHeader(query querySrv.Query, settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+func ApplyContentTypeHeader(_ context.Context, query models.Query, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
 	if strings.ToUpper(query.URLOptions.Method) == http.MethodPost {
 		switch query.URLOptions.BodyType {
 		case "raw":
@@ -66,23 +70,26 @@ func ApplyContentTypeHeader(query querySrv.Query, settings settingsSrv.InfinityS
 	return req
 }
 
-func ApplyHeadersFromSettings(settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+func ApplyAcceptEncodingHeader(_ context.Context, query models.Query, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+	req.Header.Set(headerKeyAcceptEncoding, "gzip")
+	return req
+}
+
+func ApplyHeadersFromSettings(_ context.Context, pCtx *backend.PluginContext, requestHeaders map[string]string, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
 	for key, value := range settings.CustomHeaders {
-		val := dummyHeader
+		headerValue := dummyHeader
 		if includeSect {
-			val = value
+			headerValue = value
 		}
+		headerValue = interpolateGrafanaMetaDataMacros(headerValue, pCtx)
 		if key != "" {
-			req.Header.Add(key, val)
-			if strings.EqualFold(key, headerKeyAccept) || strings.EqualFold(key, headerKeyContentType) {
-				req.Header.Set(key, val)
-			}
+			req.Header.Set(key, headerValue)
 		}
 	}
 	return req
 }
 
-func ApplyHeadersFromQuery(query querySrv.Query, settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+func ApplyHeadersFromQuery(_ context.Context, query models.Query, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
 	for _, header := range query.URLOptions.Headers {
 		value := dummyHeader
 		if includeSect {
@@ -98,30 +105,30 @@ func ApplyHeadersFromQuery(query querySrv.Query, settings settingsSrv.InfinitySe
 	return req
 }
 
-func ApplyBasicAuth(settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+func ApplyBasicAuth(_ context.Context, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
 	if settings.BasicAuthEnabled && (settings.UserName != "" || settings.Password != "") {
 		basicAuthHeader := fmt.Sprintf("Basic %s", dummyHeader)
 		if includeSect {
 			basicAuthHeader = "Basic " + base64.StdEncoding.EncodeToString([]byte(settings.UserName+":"+settings.Password))
 		}
-		req.Header.Set(headerKeyAuthorization, basicAuthHeader)
+		req.Header.Set(HeaderKeyAuthorization, basicAuthHeader)
 	}
 	return req
 }
 
-func ApplyBearerToken(settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
-	if settings.AuthenticationMethod == settingsSrv.AuthenticationMethodBearerToken {
+func ApplyBearerToken(_ context.Context, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+	if settings.AuthenticationMethod == models.AuthenticationMethodBearerToken {
 		bearerAuthHeader := fmt.Sprintf("Bearer %s", dummyHeader)
 		if includeSect {
 			bearerAuthHeader = fmt.Sprintf("Bearer %s", settings.BearerToken)
 		}
-		req.Header.Add(headerKeyAuthorization, bearerAuthHeader)
+		req.Header.Add(HeaderKeyAuthorization, bearerAuthHeader)
 	}
 	return req
 }
 
-func ApplyApiKeyAuth(settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
-	if settings.AuthenticationMethod == settingsSrv.AuthenticationMethodApiKey && settings.ApiKeyType == settingsSrv.ApiKeyTypeHeader {
+func ApplyApiKeyAuth(_ context.Context, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+	if settings.AuthenticationMethod == models.AuthenticationMethodApiKey && settings.ApiKeyType == models.ApiKeyTypeHeader {
 		apiKeyHeader := dummyHeader
 		if includeSect {
 			apiKeyHeader = settings.ApiKeyValue
@@ -133,18 +140,38 @@ func ApplyApiKeyAuth(settings settingsSrv.InfinitySettings, req *http.Request, i
 	return req
 }
 
-func ApplyForwardedOAuthIdentity(requestHeaders map[string]string, settings settingsSrv.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
+func ApplyForwardedOAuthIdentity(_ context.Context, requestHeaders map[string]string, settings models.InfinitySettings, req *http.Request, includeSect bool) *http.Request {
 	if settings.ForwardOauthIdentity {
 		authHeader := dummyHeader
 		token := dummyHeader
 		if includeSect {
-			authHeader = requestHeaders[headerKeyAuthorization]
-			token = requestHeaders[headerKeyIdToken]
+			authHeader = getQueryReqHeader(requestHeaders, HeaderKeyAuthorization)
+			token = getQueryReqHeader(requestHeaders, HeaderKeyIdToken)
 		}
-		req.Header.Add(headerKeyAuthorization, authHeader)
-		if requestHeaders[headerKeyIdToken] != "" {
-			req.Header.Add(headerKeyIdToken, token)
+		req.Header.Add(HeaderKeyAuthorization, authHeader)
+		if token != "" && token != dummyHeader {
+			req.Header.Add(HeaderKeyIdToken, token)
 		}
 	}
 	return req
+}
+
+// ApplyTraceHead injecting trace
+// https://opentelemetry.io/docs/specs/otel/context/api-propagators/#textmap-propagator
+func ApplyTraceHead(ctx context.Context, req *http.Request) *http.Request {
+	prop := otel.GetTextMapPropagator()
+	if prop != nil {
+		prop.Inject(ctx, propagation.HeaderCarrier(req.Header))
+	}
+	return req
+}
+
+func getQueryReqHeader(requestHeaders map[string]string, headerName string) string {
+	for name, value := range requestHeaders {
+		if strings.EqualFold(headerName, name) {
+			return value
+		}
+	}
+
+	return ""
 }
