@@ -1,66 +1,62 @@
 package infinity
 
 import (
+	"context"
 	"errors"
 	"fmt"
-	"strings"
 
+	"github.com/grafana/grafana-infinity-datasource/pkg/models"
+	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/yesoreyeram/grafana-framer/jsonFramer"
-	querySrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/query"
+	"github.com/grafana/infinity-libs/lib/go/jsonframer"
 )
 
-func GetFrameForInlineSources(query querySrv.Query) (*data.Frame, error) {
+func GetFrameForInlineSources(ctx context.Context, query models.Query) (*data.Frame, error) {
 	frame := GetDummyFrame(query)
-	if query.Type == querySrv.QueryTypeGROQ || query.Type == querySrv.QueryTypeUQL {
+	if query.Type == models.QueryTypeGROQ || query.Type == models.QueryTypeUQL {
 		return frame, nil
 	}
 	if query.Parser != "backend" {
 		return frame, nil
 	}
 	switch query.Type {
-	case querySrv.QueryTypeCSV, querySrv.QueryTypeTSV:
-		return GetCSVBackendResponse(query.Data, query)
-	case querySrv.QueryTypeXML, querySrv.QueryTypeHTML:
-		return GetXMLBackendResponse(query.Data, query)
-	case querySrv.QueryTypeJSON, querySrv.QueryTypeGraphQL:
-		columns := []jsonFramer.ColumnSelector{}
+	case models.QueryTypeCSV, models.QueryTypeTSV:
+		frame, err := GetCSVBackendResponse(ctx, query.Data, query)
+		if err != nil {
+			return frame, err
+		}
+		return PostProcessFrame(ctx, frame, query)
+	case models.QueryTypeXML, models.QueryTypeHTML:
+		frame, err := GetXMLBackendResponse(ctx, query.Data, query)
+		if err != nil {
+			return frame, err
+		}
+		return PostProcessFrame(ctx, frame, query)
+	case models.QueryTypeJSON, models.QueryTypeGraphQL:
+		columns := []jsonframer.ColumnSelector{}
 		for _, c := range query.Columns {
-			columns = append(columns, jsonFramer.ColumnSelector{
+			columns = append(columns, jsonframer.ColumnSelector{
 				Selector:   c.Selector,
 				Alias:      c.Text,
 				Type:       c.Type,
 				TimeFormat: c.TimeStampFormat,
 			})
 		}
-		newFrame, err := jsonFramer.JsonStringToFrame(query.Data, jsonFramer.JSONFramerOptions{
+		newFrame, err := jsonframer.ToFrame(query.Data, jsonframer.FramerOptions{
 			FrameName:    query.RefID,
 			RootSelector: query.RootSelector,
 			Columns:      columns,
 		})
 		if err != nil {
+			if errors.Is(err, jsonframer.ErrInvalidRootSelector) || errors.Is(err, jsonframer.ErrInvalidJSONContent) || errors.Is(err, jsonframer.ErrEvaluatingJSONata) {
+			return frame, backend.DownstreamError(fmt.Errorf("error converting json data to frame: %w", err))
+		}
 			return frame, err
 		}
 		if newFrame != nil {
 			frame.Fields = append(frame.Fields, newFrame.Fields...)
 		}
-		frame, err = GetFrameWithComputedColumns(frame, query.ComputedColumns)
-		if err != nil {
-			return frame, fmt.Errorf("error getting computed column. %w", err)
-		}
-		frame, err = ApplyFilter(frame, query.FilterExpression)
-		if err != nil {
-			return frame, fmt.Errorf("error applying filter. %w", err)
-		}
-		if strings.TrimSpace(query.SummarizeExpression) != "" {
-			return GetSummaryFrame(frame, query.SummarizeExpression, query.SummarizeBy)
-		}
-		if query.Format == "timeseries" && frame.TimeSeriesSchema().Type == data.TimeSeriesTypeLong {
-			if wFrame, err := data.LongToWide(frame, &data.FillMissing{Mode: data.FillModeNull}); err == nil {
-				return wFrame, err
-			}
-		}
-		return frame, nil
+		return PostProcessFrame(ctx, frame, query)
 	default:
 		return frame, errors.New("unknown backend query type")
 	}
