@@ -1,67 +1,39 @@
 package infinity
 
 import (
-	"fmt"
-	"strings"
+	"context"
 
-	xj "github.com/basgys/goxml2json"
+	"github.com/grafana/grafana-infinity-datasource/pkg/models"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/tracing"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/yesoreyeram/grafana-framer/jsonFramer"
-	querySrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/query"
+	"github.com/grafana/infinity-libs/lib/go/jsonframer"
+	"github.com/grafana/infinity-libs/lib/go/xmlframer"
 )
 
-func GetXMLBackendResponse(inputString string, query querySrv.Query) (*data.Frame, error) {
+func GetXMLBackendResponse(ctx context.Context, inputString string, query models.Query) (*data.Frame, error) {
+	_, span := tracing.DefaultTracer().Start(ctx, "GetXMLBackendResponse")
+	defer span.End()
 	frame := GetDummyFrame(query)
-	columns := []jsonFramer.ColumnSelector{}
+	columns := []jsonframer.ColumnSelector{}
 	for _, c := range query.Columns {
-		columns = append(columns, jsonFramer.ColumnSelector{
+		columns = append(columns, jsonframer.ColumnSelector{
 			Selector:   c.Selector,
 			Alias:      c.Text,
 			Type:       c.Type,
 			TimeFormat: c.TimeStampFormat,
 		})
 	}
-	xml := strings.NewReader(inputString)
-	jsonValue, err := xj.Convert(xml)
-	if err != nil {
-		backend.Logger.Error("error converting XML to frame", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
-		return frame, err
-	}
-	newFrame, err := jsonFramer.JsonStringToFrame(jsonValue.String(), jsonFramer.JSONFramerOptions{
+	newFrame, err := xmlframer.ToFrame(inputString, xmlframer.FramerOptions{
 		FrameName:    query.RefID,
 		RootSelector: query.RootSelector,
 		Columns:      columns,
 	})
-	frame.Meta = &data.FrameMeta{Custom: &CustomMeta{Query: query}}
-	if err != nil {
-		backend.Logger.Error("error getting response for query", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
-		return frame, err
-	}
 	if newFrame != nil {
 		frame.Fields = append(frame.Fields, newFrame.Fields...)
 	}
-	frame, err = GetFrameWithComputedColumns(frame, query.ComputedColumns)
 	if err != nil {
-		backend.Logger.Error("error getting computed column", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
-		return frame, err
+		err = backend.PluginError(err)
 	}
-	frame, err = ApplyFilter(frame, query.FilterExpression)
-	if err != nil {
-		backend.Logger.Error("error applying filter", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
-		return frame, fmt.Errorf("error applying filter. %w", err)
-	}
-	if strings.TrimSpace(query.SummarizeExpression) != "" {
-		return GetSummaryFrame(frame, query.SummarizeExpression, query.SummarizeBy)
-	}
-	if query.Format == "timeseries" && frame.TimeSeriesSchema().Type == data.TimeSeriesTypeLong {
-		if wFrame, err := data.LongToWide(frame, &data.FillMissing{Mode: data.FillModeNull}); err == nil {
-			return wFrame, err
-		}
-	}
-	return frame, nil
+	return frame, err
 }
