@@ -1,7 +1,7 @@
 import { evaluate, parse } from 'groq-js';
 import { sendAsDataFrame } from './UQLProvider';
 import type { InfinityQueryFormat } from './../types';
-import type { DataFrame } from '@grafana/data/types';
+import type { DataFrame } from '@grafana/data';
 
 export const getGroqResults = async (query = '*', data: unknown): Promise<any> => {
   let groqQuery = query;
