@@ -1,10 +1,9 @@
-import { toDataFrame, FieldColorModeId } from '@grafana/data';
+import { toDataFrame, FieldColorModeId, DataFrame, PreferredVisualisationType } from '@grafana/data';
 import { flatten, uniq } from 'lodash';
 import { isDataQuery } from './../utils';
 import { filterResults } from './filter';
 import { normalizeColumns } from './utils';
 import type { GrafanaTableRow, InfinityColumn, InfinityQuery, tableResult, timeSeriesResult } from './../../types';
-import type { DataFrame, PreferredVisualisationType } from '@grafana/data/types';
 
 export class InfinityParser<T extends InfinityQuery> {
   target: T;
