@@ -1,11 +1,11 @@
 import { css } from '@emotion/css';
 import React, { ComponentProps } from 'react';
-import { stylesFactory, useTheme2, ReactUtils, Field, Icon, PopoverContent, Tooltip, Tag } from '@grafana/ui';
-import { Space } from './Space';
-import { Stack } from './Stack';
+import { stylesFactory, useTheme2, ReactUtils, Field, Icon, PopoverContent, Tooltip, Tag, Stack, Space } from '@grafana/ui';
 import type { GrafanaTheme2 } from '@grafana/data';
 
-export const EditorFieldGroup: React.FC<{}> = ({ children }) => {
+export const EditorFieldGroup: React.FC<{
+  children: React.ReactNode;
+}> = ({ children }) => {
   return <Stack gap={1}>{children}</Stack>;
 };
 
@@ -20,17 +20,18 @@ interface EditorFieldProps extends ComponentProps<typeof Field> {
   tag?: string;
   dataTestId?: string;
   borderColor?: string;
+  horizontal?: boolean;
 }
 
 export const EditorField: React.FC<EditorFieldProps> = (props) => {
-  const { label, optional, tooltip, children, promoNode, width, invalid, borderColor, tag, dataTestId, ...fieldProps } = props;
+  const { label, optional, tooltip, children, promoNode, width, invalid, borderColor, tag, dataTestId, horizontal, ...fieldProps } = props;
 
   const theme = useTheme2();
-  const styles = getStyles(theme, width, invalid ? 'red' : borderColor);
+  const styles = getStyles(theme, width, invalid ? 'red' : borderColor, horizontal);
 
   // Null check for backward compatibility
   const childInputId = fieldProps?.htmlFor || ReactUtils?.getChildId(children);
-  const testId = (compType = '') => `infinity-query-field${compType ? '-' + compType : ''}-${(dataTestId || label).replace(/\ /g, '-')}`.toLowerCase();
+  const testId = (compType = '') => `infinity-query-field${compType ? '-' + compType : ''}-${(dataTestId || label).replace(/ /g, '-')}`.toLowerCase();
 
   const labelEl = (
     <>
@@ -49,6 +50,16 @@ export const EditorField: React.FC<EditorFieldProps> = (props) => {
     </>
   );
 
+  if (horizontal) {
+    return (
+      <div className={styles.root} data-testid={testId('wrapper')}>
+        <Field className={styles.field} label={labelEl} {...fieldProps} horizontal={true}>
+          {children}
+        </Field>
+      </div>
+    );
+  }
+
   return (
     <div className={styles.root} data-testid={testId('wrapper')}>
       <Field className={styles.field} label={labelEl} {...fieldProps}>
@@ -58,19 +69,23 @@ export const EditorField: React.FC<EditorFieldProps> = (props) => {
   );
 };
 
-const getStyles = stylesFactory((theme: GrafanaTheme2, width?: number | string, borderColor = 'transparent') => {
+const getStyles = stylesFactory((theme: GrafanaTheme2, width?: number | string, borderColor = 'transparent', horizontal = false) => {
   return {
     root: css({
       minWidth: theme.spacing(width ?? 0),
-      paddingInlineStart: '7px',
-      paddingInlineEnd: '5px',
-      borderLeft: `1px solid ${borderColor}`,
-      marginRight: '5px',
+      // boxShadow: `0px 0px 4px 0px ${theme.colors.border.weak}`,
+      border: `1px solid ${theme.colors.border.medium}`,
+      padding: theme.spacing('8px'),
+      background: theme.colors.background.primary,
+      // marginRight: horizontal ? '10px' : '5px',
     }),
     label: css({
-      fontSize: 12,
+      fontSize: theme.typography.bodySmall.fontSize,
       fontWeight: theme.typography.fontWeightMedium,
       paddingLeft: '1px',
+      border: horizontal ? `1px solid ${borderColor}` : '',
+      padding: horizontal ? '5px 10px 5px 0px' : '',
+      textAlign: horizontal ? 'right' : 'left',
     }),
     tag: css({
       marginLeft: '10px',
