@@ -1,8 +1,7 @@
 import { css } from '@emotion/css';
 import React, { useState } from 'react';
 import { GrafanaTheme2 } from '@grafana/data';
-import { useStyles2, Button } from '@grafana/ui';
-import { Stack } from './Stack';
+import { useStyles2, Button, Stack, useTheme2 } from '@grafana/ui';
 
 interface EditorRowProps {
   label: string;
@@ -10,16 +9,18 @@ interface EditorRowProps {
   title?: () => string | React.ReactNode;
   collapsed?: boolean;
   dataTestId?: string;
+  children: React.ReactNode;
 }
 
 export const EditorRow: React.FC<EditorRowProps> = ({ label, collapsible, collapsed = true, title, dataTestId, children }) => {
   const styles = useStyles2(getStyles);
+  const theme = useTheme2();
   const [show, setShow] = useState(collapsed);
-  const testId = (compType = '') => `infinity-query-row${compType ? '-' + compType : ''}-${(dataTestId || label).replace(/\ /g, '-')}`.toLowerCase();
+  const testId = (compType = '') => `infinity-query-row${compType ? '-' + compType : ''}-${(dataTestId || label).replace(/ /g, '-')}`.toLowerCase();
   return (
     <div className={styles.root} data-testid={testId('wrapper')}>
       {collapsible && (
-        <>
+        <div style={{ display: 'flex' }}>
           <Button
             icon={show ? 'angle-down' : 'angle-right'}
             fill="text"
@@ -42,10 +43,24 @@ export const EditorRow: React.FC<EditorRowProps> = ({ label, collapsible, collap
             <b className={styles.collapseTile}>{label}</b>
           </span>
           <span className={styles.collapseTileSecondary}>{title ? title() : 'Options'}</span>
-        </>
+        </div>
       )}
       {show && (
-        <div style={{ display: 'flex', marginTop: label && collapsible ? '15px' : '0px', marginLeft: '0px' }} data-testid={testId(`children`)}>
+        <div
+          style={{
+            display: 'flex',
+            flexWrap: 'wrap',
+            justifyContent: 'flex-start',
+            alignContent: 'flex-start',
+            alignItems: 'flex-start',
+            gap: theme.spacing(1),
+            marginTop: label && collapsible ? '15px' : '0px',
+            marginLeft: '0px',
+            flexDirection: 'row',
+            width: '100%',
+          }}
+          data-testid={testId(`children`)}
+        >
           {children}
         </div>
       )}
@@ -74,7 +89,9 @@ const getStyles = (theme: GrafanaTheme2) => {
   };
 };
 
-interface EditorRowsProps {}
+interface EditorRowsProps {
+  children: React.ReactNode;
+}
 
 export const EditorRows: React.FC<EditorRowsProps> = ({ children }) => {
   return (
