@@ -3,8 +3,7 @@ import React from 'react';
 import { useToggle } from 'react-use';
 
 import { GrafanaTheme2 } from '@grafana/data';
-import { Icon, useStyles2 } from '@grafana/ui';
-import { Stack } from './Stack';
+import { Icon, useStyles2, Stack } from '@grafana/ui';
 
 export interface Props {
   title: string;
