@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { FileDropzone, LinkButton, Button, InlineFormLabel, Input } from '@grafana/ui';
-import type { DataSourceSettings } from '@grafana/data/types';
+import type { DataSourceSettings } from '@grafana/data';
 import type { InfinityOptions } from './../../../types';
 
 type GoogleJWTEditorProps = {
