@@ -7,7 +7,7 @@ export interface GlobalInfinityQuery {
   id: string;
   query: InfinityQuery;
 }
-export type AuthType = 'none' | 'basicAuth' | 'apiKey' | 'bearerToken' | 'oauthPassThru' | 'digestAuth' | 'oauth2';
+export type AuthType = 'none' | 'basicAuth' | 'apiKey' | 'bearerToken' | 'oauthPassThru' | 'digestAuth' | 'aws' | 'azureBlob' | 'oauth2';
 export type OAuth2Type = 'client_credentials' | 'jwt' | 'others';
 export type APIKeyType = 'header' | 'query';
 export type OAuth2Props = {
@@ -18,26 +18,40 @@ export type OAuth2Props = {
   subject?: string;
   token_url?: string;
   scopes?: string[];
+  authStyle?: number;
+};
+export type AWSAuthProps = {
+  authType?: 'keys';
+  region?: string;
+  service?: string;
 };
 export type InfinityReferenceData = { name: string; data: string };
+export type ProxyType = 'none' | 'env' | 'url';
+export type UnsecureQueryHandling = 'warn' | 'allow' | 'deny';
 export interface InfinityOptions extends DataSourceJsonData {
   auth_method?: AuthType;
   apiKeyKey?: string;
   apiKeyType?: APIKeyType;
   oauth2?: OAuth2Props;
+  aws?: AWSAuthProps;
   tlsSkipVerify?: boolean;
   tlsAuth?: boolean;
   serverName?: string;
   tlsAuthWithCACert?: boolean;
   global_queries?: GlobalInfinityQuery[];
   timeoutInSeconds?: number;
+  proxy_type?: ProxyType;
+  proxy_url?: string;
   oauthPassThru?: boolean;
   allowedHosts?: string[];
   refData?: InfinityReferenceData[];
-  enableOpenApi?: boolean;
-  openApiVersion?: 'open-api-2.0';
-  openApiUrl?: string;
-  openAPIBaseURL?: string;
+  customHealthCheckEnabled?: boolean;
+  customHealthCheckUrl?: string;
+  azureBlobAccountUrl?: string;
+  azureBlobAccountName?: string;
+  unsecuredQueryHandling?: UnsecureQueryHandling;
+  enableSecureSocksProxy?: boolean;
+  pathEncodedUrlsEnabled?: boolean;
 }
 
 export interface InfinitySecureOptions {
@@ -47,8 +61,11 @@ export interface InfinitySecureOptions {
   tlsClientKey?: string;
   apiKeyValue?: string;
   bearerToken?: string;
+  awsAccessKey?: string;
+  awsSecretKey?: string;
   oauth2ClientSecret?: string;
   oauth2JWTPrivateKey?: string;
+  azureBlobAccountKey?: string;
 }
 export interface SecureField {
   id: string;
