@@ -2,10 +2,10 @@ import { FilterOperator } from './../constants';
 import type { DataQuery, SelectableValue } from '@grafana/data';
 
 //#region Query
-export type InfinityQueryType = 'json' | 'csv' | 'tsv' | 'xml' | 'graphql' | 'html' | 'series' | 'global' | 'uql' | 'groq' | 'google-sheets';
-export type InfinityQuerySources = 'url' | 'inline' | 'reference' | 'random-walk' | 'expression';
-export type InfinityColumnFormat = 'string' | 'number' | 'timestamp' | 'timestamp_epoch' | 'timestamp_epoch_s';
-export type InfinityQueryFormat = 'table' | 'timeseries' | 'dataframe' | 'as-is' | 'node-graph-nodes' | 'node-graph-edges';
+export type InfinityQueryType = 'json' | 'csv' | 'tsv' | 'xml' | 'graphql' | 'html' | 'series' | 'global' | 'uql' | 'groq' | 'google-sheets' | 'transformations';
+export type InfinityQuerySources = 'url' | 'inline' | 'azure-blob' | 'reference' | 'random-walk' | 'expression';
+export type InfinityColumnFormat = 'string' | 'number' | 'timestamp' | 'timestamp_epoch' | 'timestamp_epoch_s' | 'boolean';
+export type InfinityQueryFormat = 'table' | 'timeseries' | 'logs' | 'trace' | 'node-graph-nodes' | 'node-graph-edges' | 'dataframe' | 'as-is';
 export type QueryBodyType = 'none' | 'form-data' | 'x-www-form-urlencoded' | 'raw' | 'graphql';
 export type QueryBodyContentType = 'text/plain' | 'application/json' | 'application/xml' | 'text/html' | 'application/javascript';
 export type InfinityQueryBase<T extends InfinityQueryType> = { type: T } & DataQuery;
@@ -20,7 +20,7 @@ export type InfinityURLOptions = {
   body_content_type?: QueryBodyContentType;
   body_form?: InfinityKV[];
   body_graphql_query?: string;
-  // body_graphql_variables?: string;
+  body_graphql_variables?: string;
 };
 export type InfinityQueryWithReferenceSource<T extends InfinityQueryType> = {
   referenceName: string;
@@ -30,6 +30,11 @@ export type InfinityQueryWithURLSource<T extends InfinityQueryType> = {
   url: string;
   url_options: InfinityURLOptions;
 } & InfinityQueryWithSource<'url'> &
+  InfinityQueryBase<T>;
+export type InfinityQueryWithAzureBlobSource<T extends InfinityQueryType> = {
+  azContainerName: string;
+  azBlobName: string;
+} & InfinityQueryWithSource<'azure-blob'> &
   InfinityQueryBase<T>;
 export type InfinityQueryWithInlineSource<T extends InfinityQueryType> = {
   data: string;
@@ -42,17 +47,16 @@ export type InfinityQueryWithDataSource<T extends InfinityQueryType> = {
   columns: InfinityColumn[];
   filters?: InfinityFilter[];
   format: InfinityQueryFormat;
-} & (InfinityQueryWithURLSource<T> | InfinityQueryWithInlineSource<T> | InfinityQueryWithReferenceSource<T>) &
+} & (InfinityQueryWithURLSource<T> | InfinityQueryWithInlineSource<T> | InfinityQueryWithReferenceSource<T> | InfinityQueryWithAzureBlobSource<T>) &
   InfinityQueryBase<T>;
 export type InfinityJSONQueryOptions = {
   root_is_not_array?: boolean;
   columnar?: boolean;
 };
-export type BackendParserOptions = { filterExpression?: string; summarizeExpression?: string; summarizeBy?: string; computed_columns?: InfinityColumn[] };
+export type BackendParserOptions = { filterExpression?: string; summarizeExpression?: string; summarizeAlias?: string; summarizeBy?: string; computed_columns?: InfinityColumn[] };
 export type InfinityJSONQuery = (
   | { parser?: 'simple'; json_options?: InfinityJSONQueryOptions }
   | ({ parser: 'backend' } & BackendParserOptions)
-  | { parser: 'sqlite'; sqlite_query?: string }
   | { parser: 'uql'; uql?: string }
   | { parser: 'groq'; groq?: string }
 ) &
@@ -99,7 +103,64 @@ export type InfinityUQLQuery = { uql: string; format: InfinityQueryFormat } & In
 export type InfinityGROQQuerySource = InfinityQueryWithURLSource<'groq'> | InfinityQueryWithInlineSource<'groq'>;
 export type InfinityGROQQuery = { groq: string; format: InfinityQueryFormat } & InfinityGROQQuerySource & InfinityQueryBase<'groq'>;
 export type InfinityGSheetsQuery = { spreadsheet: string; sheetName?: string; range: string; columns: InfinityColumn[] } & InfinityQueryBase<'google-sheets'>;
-export type InfinityQuery = InfinityLegacyQuery | InfinityUQLQuery | InfinityGROQQuery | InfinityGSheetsQuery;
+export type PaginationType = 'none' | 'offset' | 'page' | 'cursor' | 'list';
+export type PaginationParamType = 'query' | 'header' | 'body_data' | 'body_json' | 'replace';
+export type PaginationBase<T extends PaginationType> = { pagination_mode?: T; pagination_max_pages?: number };
+export type PaginationNone = {} & PaginationBase<'none'>;
+export type PaginationOffset = {
+  pagination_param_size_field_name?: string;
+  pagination_param_size_field_type?: PaginationParamType;
+  pagination_param_size_value?: number;
+  pagination_param_offset_field_name?: string;
+  pagination_param_offset_field_type?: PaginationParamType;
+  pagination_param_offset_value?: number;
+} & PaginationBase<'offset'>;
+export type PaginationPage = {
+  pagination_param_size_field_name?: string;
+  pagination_param_size_field_type?: PaginationParamType;
+  pagination_param_size_value?: number;
+  pagination_param_page_field_name?: string;
+  pagination_param_page_field_type?: PaginationParamType;
+  pagination_param_page_value?: number;
+} & PaginationBase<'page'>;
+export type PaginationCursor = {
+  pagination_param_size_field_name?: string;
+  pagination_param_size_field_type?: PaginationParamType;
+  pagination_param_size_value?: number;
+  pagination_param_cursor_field_name?: string;
+  pagination_param_cursor_field_type?: PaginationParamType;
+  pagination_param_cursor_extraction_path?: string;
+} & PaginationBase<'cursor'>;
+export type PaginationList = {
+  pagination_param_list_field_name?: string;
+  pagination_param_list_field_type?: PaginationParamType;
+  pagination_param_list_value?: string;
+} & PaginationBase<'list'>;
+export type Pagination = PaginationNone | PaginationOffset | PaginationPage | PaginationCursor | PaginationList;
+export type Transformation = 'limit' | 'filterExpression' | 'summarize' | 'computedColumn';
+export type TransformationItem = {
+  type: Transformation;
+  disabled?: boolean;
+  limit?: {
+    limitField?: number;
+  };
+  filterExpression?: {
+    expression?: string;
+  };
+  computedColumn?: {
+    expression?: string;
+    alias: string;
+  };
+  summarize?: {
+    expression?: string;
+    by?: string;
+    alias?: string;
+  };
+};
+export type TransformationsQuery = {
+  transformations: TransformationItem[];
+} & InfinityQueryBase<'transformations'>;
+export type InfinityQuery = (InfinityLegacyQuery | InfinityUQLQuery | InfinityGROQQuery | InfinityGSheetsQuery | TransformationsQuery) & Pagination;
 //#endregion
 
 //#region Misc
